+++ src/titiler/application/titiler/application/__init__.py
@@ -1,3 +1,3 @@
 """titiler.application"""
 
-__version__ = "0.25.0"
+__version__ = "0.26.0"